Issue #560: [extras-testing-tools] Add Scala.js support - Move StubTools to the jvm specific folder
