+++ modules/extras-testing-tools/jvm/src/test/scala/extras/testing/TestType.scala
